feat(vault): add cfxOnly to public vault
if user giving a base32 address vault, mark it as cfx only vault, don't create
ethHexAddress on the corresponding account when creating it from the vault

For hex address public vault, store the original hex address. Create a normal
account with ethHexAddress and convert it to 0x-prefixed address as cfxHexAddress.